Implements review comments
