+++ lib/modules/platform/bitbucket-server/utils.ts
@@ -6,6 +6,7 @@ import { logger } from '../../../logger';
 import type { HostRule } from '../../../types';
 import type { GitOptions, GitProtocol } from '../../../types/git';
 import * as git from '../../../util/git';
+import { regEx } from '../../../util/regex';
 import { parseUrl } from '../../../util/url';
 import { getPrBodyStruct } from '../pr-body';
 import type { GitUrlOption } from '../types';
@@ -147,4 +148,29 @@ export function getExtraCloneOpts(opts: HostRule): GitOptions {
     };
   }
   return {};
+}
+
+export function splitEscapedSpaces(str: string): string[] {
+  const parts = str.split(' ');
+  const result: string[] = [];
+  let last: string | undefined;
+
+  for (const part of parts) {
+    if (last?.endsWith(String.raw`\\`)) {
+      result[result.length - 1] = last.slice(0, -2) + ' ' + part;
+    } else {
+      result.push(part);
+    }
+    last = result[result.length - 1];
+  }
+
+  return result;
+}
+
+export function parseModifier(value: string): number | null {
+  const match = regEx('^random(?:\\((\\d+)\\))?$').exec(value);
+  if (!match) {
+    return null;
+  }
+  return parseInt(match[1] ?? '1', 10);
 }
